bot/src/requirements.txt
discord.py>=2.3.2
sqlalchemy>=2.0.0
asyncpg>=0.29.0
python-dotenv>=1.0.0
aiohttp>=3.9.1
psycopg2-binary>=2.9.9
alembic>=1.13.1
pydantic>=2.5.3
uvicorn>=0.25.0
fastapi>=0.109.0
python-jose>=3.3.0
passlib>=1.7.4
bcrypt>=4.1.2
python-multipart>=0.0.6
aiofiles>=23.2.1
pillow>=10.2.0
python-magic>=0.4.27
regex>=2023.12.25
numpy>=1.26.3
pandas>=2.2.0
matplotlib>=3.8.2
seaborn>=0.13.1
requests>=2.31.0
beautifulsoup4>=4.12.3
lxml>=5.1.0
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
black>=24.1.1
flake8>=7.0.0
mypy>=1.8.0
isort>=5.13.2
pre-commit>=3.6.0
google-cloud-translate>=3.12.0 